Add `Binding.removeDuplicates()`
Because this library makes it easy to add logic around navigation, it's
probably also a good idea to ship with helpers that work around some
surprising bugs/behaviors that currently exist in SwiftUI. For example,
as noticed by #10, `NavigationLink` writes `nil` to its binding twice on
dismissal.

It's probably not appropriate for us to automatically filter duplicate
writes, but we can at least ship a `Binding.removeDuplicates()` that
makes it easy to achieve this behavior.