Fix `AlertState.init` availability (#198)
In SwiftUI Navigation 1.0, `AlertState` had many initializers to match
the various interfaces in SwiftUI, and they matched the availability of
the SwiftUI interfaces. We simplified to a single interface in 2.0, but
didn't consider the availability mismatch.

While we could bring back the old APIs, it might be nice to favor the
modern builder style. The one caveat is that it will be possible to
render alerts pre-iOS 15 that are configured with more than the maximum
2 buttons at that time. We can emit a runtime warning if we detect this,
but we are losing some of the compile-time check of the previous
version.