Support eager navigation destination resolution
Currently we steer folks to using `viewDidLoad` for all observation, but
this laziness can lead to issues with deep-linking in type-erased
navigation stacks, where a destination may declare a sub-destination in
its `viewDidLoad`, causing the first drill-down to occur lazily over
multiple animated steps instead of a single one. This behavior can be
seen in the type-erased navigation case study in the repository.

This PR is a proof of concept to show that the library can populate a
"current navigation stack" in the UI transaction so that a destination's
initializer can declare its dependent destinations, fixing the case
study behavior.