Track VCs being popped from stack from UIKit
Tapping the back button or interactively popping via gesture can
conflict with observation when the view controller being popped to
executes a mutation in its `will`- or `didAppear`. While this isn't
common in vanilla UIKit, in TCA sending even a no-op action in these
lifecycle hooks can lead to immediately re-pushing the view controller
that was just popped.

This commit does what it can to detect when the back button was pushed
or the interactive pop gesture was invoked so that if/when the pop is
committed, we can update the path more eagerly, avoiding the
re-presentation.