Add UIBinding deprecation warnings to avoid misuse (#284)
* Add UIBinding deprecation warnings to avoid misuse

The `@UIBinding` property wrapper is meant as a substitute for `@State`,
where you can introduce local, owned state to a view controller that can
be strongly held in another child, but because it is completely
unconstrained it was easy to reach for instead of `@UIBindable` when it
came to view models, and those view models could be retained too
strongly.

This PR introduces a couple deprecated overloads to catch this misuse
and guide folks towards `@UIBindable`, instead.

We might want to consider this subtlety a bit more and rethink or better
document the differences.

* Disable Windows CI

* wip